Cache the entire `tmp/${{ ruby-platform }}`
`tmp/${{ ruby-platform }}/target` does not exist. On wasmtime-rb, the
folder structure is `tmp/{platform}/wasmtime_rb/{ruby-version}`.
Here's what I get when I run `tree -d` with a certain depth limit[1]:

```
tmp
└── x86_64-linux
    ├── stage
    │   ├── ext
    │   │   └── src
    │   │       ├── helpers
    │   │       └── ruby_api
    │   └── lib
    │       └── wasmtime
    └── wasmtime_rb
        └── 3.1.3
            └── target
                └── release
                    ├── build
                    ├── deps
                    ├── examples
                    └── incremental
```
This will cache the `stage` folder -- I think it's fine, but unsure.

Also adds the ruby version to the cache key because the extension
is compiled in a version-specific folder.

[1]: https://github.com/bytecodealliance/wasmtime-rb/actions/runs/3824179901/jobs/6506092845