Add GH Action to test cross-compilation
Strictly speaking, cross-compiling with rb-sys isn't foolproof; no
software is. This commit introduces a new GH Action for consumers to
make use of. It simply untars the cross-compiled gem and asserts that
files which are supposed to be there, such as  .{so,bundle} in
`lib/$RUBY_VERSION` are indeed there